requirements.txt
# Add specific versions if known or required
google-cloud-aiplatform[adk,agent-engines]==1.93.0
google-adk
cloudpickle==3.1.1
crewai
crewai-tools
pydantic==2.11.3